requirements.txt
sentence-transformers
chromadb
google-generativeai
python-dotenv
numpy
